Implement Restaurant Index Action
Worked with Dan and Wei.
Created restaurant 'index' script.
Added Restaurants route for 'index'
Updated restaurants controller to require the owner to match to include
restaurant in GET request.
Tested with script - feature is working and restaurants are only shown
by user ID.
Next will be to delete a restaurant.